Make delete space idempotent

+++ src/cf/commands/delete_space.go
@@ -23,32 +23,23 @@ func NewDeleteSpace(ui terminal.UI, sR api.SpaceRepository) (cmd *DeleteSpace) {
 }
 
 func (cmd *DeleteSpace) GetRequirements(reqFactory requirements.Factory, c *cli.Context) (reqs []requirements.Requirement, err error) {
-	var spaceName string
-
-	if len(c.Args()) == 1 {
-		spaceName = c.Args()[0]
-	}
-
-	if spaceName == "" {
+	if len(c.Args()) != 1 {
 		err = errors.New("Incorrect Usage")
 		cmd.ui.FailWithUsage(c, "delete-space")
 		return
 	}
 
-	cmd.spaceReq = reqFactory.NewSpaceRequirement(spaceName)
-
-	reqs = []requirements.Requirement{cmd.spaceReq}
 	return
 }
 
 func (cmd *DeleteSpace) Run(c *cli.Context) {
-	space := cmd.spaceReq.GetSpace()
-	force := c.Bool("f")
+	spaceName := c.Args()[0]
 
+	force := c.Bool("f")
 	if !force {
 		response := strings.ToLower(cmd.ui.Ask(
 			"Really delete space %s and everything associated with it?%s",
-			terminal.EntityNameColor(space.Name),
+			terminal.EntityNameColor(spaceName),
 			terminal.PromptColor(">"),
 		))
 		if response != "y" && response != "yes" {
@@ -56,7 +47,21 @@ func (cmd *DeleteSpace) Run(c *cli.Context) {
 		}
 	}
 
-	cmd.ui.Say("Deleting space %s...", terminal.EntityNameColor(space.Name))
+	cmd.ui.Say("Deleting space %s...", terminal.EntityNameColor(spaceName))
+
+	space, found, apiErr := cmd.spaceRepo.FindByName(spaceName)
+
+	if apiErr != nil {
+		cmd.ui.Failed(apiErr.Error())
+		return
+	}
+
+	if !found {
+		cmd.ui.Ok()
+		cmd.ui.Say("Space %s was already deleted.", terminal.EntityNameColor(spaceName))
+		return
+	}
+
 	err := cmd.spaceRepo.Delete(space)
 	if err != nil {
 		cmd.ui.Failed(err.Error())
